+++ nautilus_core/model/src/data/tick.rs
@@ -17,7 +17,7 @@ use std::fmt::{Display, Formatter, Result};
 
 use pyo3::ffi;
 
-use crate::enums::OrderSide;
+use crate::enums::AggressorSide;
 use crate::identifiers::instrument_id::InstrumentId;
 use crate::identifiers::trade_id::TradeId;
 use crate::types::price::Price;
@@ -90,7 +90,7 @@ pub struct TradeTick {
     pub instrument_id: InstrumentId,
     pub price: Price,
     pub size: Quantity,
-    pub aggressor_side: OrderSide,
+    pub aggressor_side: AggressorSide,
     pub trade_id: TradeId,
     pub ts_event: Timestamp,
     pub ts_init: Timestamp,
@@ -101,7 +101,7 @@ impl TradeTick {
         instrument_id: InstrumentId,
         price: Price,
         size: Quantity,
-        aggressor_side: OrderSide,
+        aggressor_side: AggressorSide,
         trade_id: TradeId,
         ts_event: Timestamp,
         ts_init: Timestamp,
@@ -139,6 +139,11 @@ impl Display for TradeTick {
 #[no_mangle]
 pub extern "C" fn quote_tick_free(tick: QuoteTick) {
     drop(tick); // Memory freed here
+}
+
+#[no_mangle]
+pub extern "C" fn quote_tick_copy(tick: &QuoteTick) -> QuoteTick {
+    tick.clone()
 }
 
 #[no_mangle]
@@ -204,13 +209,18 @@ pub extern "C" fn trade_tick_free(tick: TradeTick) {
 }
 
 #[no_mangle]
+pub extern "C" fn trade_tick_copy(tick: &TradeTick) -> TradeTick {
+    tick.clone()
+}
+
+#[no_mangle]
 pub extern "C" fn trade_tick_from_raw(
     instrument_id: InstrumentId,
     price: i64,
     price_prec: u8,
     size: u64,
     size_prec: u8,
-    aggressor_side: OrderSide,
+    aggressor_side: AggressorSide,
     trade_id: TradeId,
     ts_event: u64,
     ts_init: u64,
@@ -243,7 +253,7 @@ pub unsafe extern "C" fn trade_tick_to_pystr(tick: &TradeTick) -> *mut ffi::PyOb
 #[cfg(test)]
 mod tests {
     use crate::data::tick::{QuoteTick, TradeTick};
-    use crate::enums::OrderSide;
+    use crate::enums::AggressorSide;
     use crate::identifiers::instrument_id::InstrumentId;
     use crate::identifiers::trade_id::TradeId;
     use crate::types::price::Price;
@@ -273,7 +283,7 @@ mod tests {
             instrument_id: InstrumentId::from("ETHUSDT-PERP.BINANCE"),
             price: Price::new(10000.0, 4),
             size: Quantity::new(1.0, 8),
-            aggressor_side: OrderSide::Buy,
+            aggressor_side: AggressorSide::Buy,
             trade_id: TradeId::new("123456789"),
             ts_event: 0,
             ts_init: 0,