+++ nautilus_core/model/src/enums.rs
@@ -13,159 +13,79 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::fmt::{Debug, Display, Formatter, Result};
+use std::fmt::Debug;
+use std::str::FromStr;
+
+use pyo3::ffi;
+use strum::{Display, EnumString, FromRepr};
+
+use nautilus_core::string::{pystr_to_string, string_to_pystr};
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-#[allow(non_camel_case_types)]
-pub enum CurrencyType {
-    Crypto = 1,
-    Fiat = 2,
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum AccountType {
+    Cash = 1,
+    Margin = 2,
+    Betting = 3,
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-#[allow(non_camel_case_types)]
-pub enum OrderSide {
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum AggregationSource {
+    External = 1,
+    Internal = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum AggressorSide {
     None = 0,
     Buy = 1,
     Sell = 2,
 }
 
-impl From<u8> for OrderSide {
-    fn from(v: u8) -> Self {
-        match v {
-            0 => OrderSide::None,
-            1 => OrderSide::Buy,
-            2 => OrderSide::Sell,
-            _ => panic!("Invalid `OrderSide` value, was {v}"),
-        }
-    }
-}
-
-impl From<&str> for OrderSide {
-    fn from(s: &str) -> Self {
-        match s.to_uppercase().as_str() {
-            "NONE" => OrderSide::None,
-            "BUY" => OrderSide::Buy,
-            "SELL" => OrderSide::Sell,
-            _ => panic!("Invalid `OrderSide` value, was {s}"),
-        }
-    }
-}
-
-impl OrderSide {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            OrderSide::None => "NONE",
-            OrderSide::Buy => "BUY",
-            OrderSide::Sell => "SELL",
-        }
-    }
-}
-
-// impl Into<&str> for OrderSide {
-//     fn into(self) -> &'static str {
-//         match self {
-//             OrderSide::Buy => "BUY",
-//             OrderSide::Sell => "SELL",
-//         }
-//     }
-// }
-
-// impl ToString for OrderSide {
-//     fn to_string(&self) -> String {
-//         match self {
-//             OrderSide::Buy => "BUY",
-//             OrderSide::Sell => "SELL",
-//         }.parse().unwrap()
-//     }
-// }
-
-impl Display for OrderSide {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.as_str())
-    }
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+#[allow(non_camel_case_types)]
+pub enum AssetClass {
+    FX = 1,
+    Equity = 2,
+    Commodity = 3,
+    Metal = 4,
+    Energy = 5,
+    Bond = 6,
+    Index = 7,
+    Crypto = 8,
+    Betting = 9,
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
 #[allow(non_camel_case_types)]
-pub enum PriceType {
-    Bid = 1,
-    Ask = 2,
-    Mid = 3,
-    Last = 4,
-}
-impl PriceType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            PriceType::Bid => "BID",
-            PriceType::Ask => "ASK",
-            PriceType::Mid => "MID",
-            PriceType::Last => "LAST",
-        }
-    }
-}
-impl Display for PriceType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.as_str())
-    }
-}
-
-impl From<&str> for PriceType {
-    fn from(s: &str) -> Self {
-        match s.to_uppercase().as_str() {
-            "BID" => PriceType::Bid,
-            "ASK" => PriceType::Ask,
-            "MID" => PriceType::Mid,
-            "LAST" => PriceType::Last,
-            _ => panic!("Invalid `PriceType` value, was {s}"),
-        }
-    }
-}
-impl From<u8> for PriceType {
-    fn from(i: u8) -> Self {
-        match i {
-            1 => PriceType::Bid,
-            2 => PriceType::Ask,
-            3 => PriceType::Mid,
-            4 => PriceType::Last,
-            _ => panic!("Invalid `Price` value, was {i}"),
-        }
-    }
+pub enum AssetType {
+    Spot = 1,
+    Swap = 2,
+    Future = 3,
+    Forward = 4,
+    Cfd = 5,
+    Option = 6,
+    Warrant = 7,
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-#[allow(non_camel_case_types)]
-pub enum BookLevel {
-    L1_TBBO = 1,
-    L2_MBP = 2,
-    L3_MBO = 3,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-#[allow(non_camel_case_types)]
-pub enum BookAction {
-    Add = 1,
-    Update = 2,
-    Delete = 3,
-    Clear = 4,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-#[allow(non_camel_case_types)]
-pub enum DepthType {
-    Volume = 1,
-    Exposure = 2,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, PartialOrd)]
-#[allow(non_camel_case_types)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
 pub enum BarAggregation {
     Tick = 1,
     TickImbalance = 2,
@@ -185,120 +105,323 @@ pub enum BarAggregation {
     Month = 16,
 }
 
-impl From<&str> for BarAggregation {
-    fn from(s: &str) -> Self {
-        match s.to_uppercase().as_str() {
-            "TICK" => BarAggregation::Tick,
-            "TICK_IMBALANCE" => BarAggregation::TickImbalance,
-            "TICK_RUNS" => BarAggregation::TickRuns,
-            "VOLUME" => BarAggregation::Volume,
-            "VOLUME_IMBALANCE" => BarAggregation::VolumeImbalance,
-            "VOLUME_RUNS" => BarAggregation::VolumeRuns,
-            "VALUE" => BarAggregation::Value,
-            "VALUE_IMBALANCE" => BarAggregation::ValueImbalance,
-            "VALUE_RUNS" => BarAggregation::ValueRuns,
-            "MILLISECOND" => BarAggregation::Millisecond,
-            "SECOND" => BarAggregation::Second,
-            "MINUTE" => BarAggregation::Minute,
-            "HOUR" => BarAggregation::Hour,
-            "DAY" => BarAggregation::Day,
-            "WEEK" => BarAggregation::Week,
-            "MONTH" => BarAggregation::Month,
-            _ => panic!("Invalid `BarAggregation` value, was {s}"),
-        }
-    }
-}
-impl From<u8> for BarAggregation {
-    fn from(i: u8) -> Self {
-        match i {
-            1 => BarAggregation::Tick,
-            2 => BarAggregation::TickImbalance,
-            3 => BarAggregation::TickRuns,
-            4 => BarAggregation::Volume,
-            5 => BarAggregation::VolumeImbalance,
-            6 => BarAggregation::VolumeRuns,
-            7 => BarAggregation::Value,
-            8 => BarAggregation::ValueImbalance,
-            9 => BarAggregation::ValueRuns,
-            10 => BarAggregation::Millisecond,
-            11 => BarAggregation::Second,
-            12 => BarAggregation::Minute,
-            13 => BarAggregation::Hour,
-            14 => BarAggregation::Day,
-            15 => BarAggregation::Week,
-            16 => BarAggregation::Month,
-            _ => panic!("Invalid `BarAggregation` value, was {i}"),
-        }
-    }
-}
-
-impl BarAggregation {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            BarAggregation::Tick => "TICK",
-            BarAggregation::TickImbalance => "TICK_IMBALANCE",
-            BarAggregation::TickRuns => "TICK_RUNS",
-            BarAggregation::Volume => "VOLUME",
-            BarAggregation::VolumeImbalance => "VOLUME_IMBALANCE",
-            BarAggregation::VolumeRuns => "VOLUME_RUNS",
-            BarAggregation::Value => "VALUE",
-            BarAggregation::ValueImbalance => "VALUE_IMBALANCE",
-            BarAggregation::ValueRuns => "VALUE_RUNS",
-            BarAggregation::Millisecond => "MILLISECOND",
-            BarAggregation::Second => "SECOND",
-            BarAggregation::Minute => "MINUTE",
-            BarAggregation::Hour => "HOUR",
-            BarAggregation::Day => "DAY",
-            BarAggregation::Week => "WEEK",
-            BarAggregation::Month => "MONTH",
-        }
-    }
-}
-
-impl Display for BarAggregation {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.as_str())
-    }
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum BookAction {
+    Add = 1,
+    Update = 2,
+    Delete = 3,
+    Clear = 4,
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
 #[allow(non_camel_case_types)]
-pub enum AggregationSource {
-    EXTERNAL = 1,
-    INTERNAL = 2,
+pub enum BookType {
+    L1_TBBO = 1,
+    L2_MBP = 2,
+    L3_MBO = 3,
 }
 
-impl AggregationSource {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AggregationSource::EXTERNAL => "EXTERNAL",
-            AggregationSource::INTERNAL => "INTERNAL",
-        }
-    }
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+#[allow(non_camel_case_types)]
+pub enum ContingencyType {
+    None = 0,
+    Oco = 1,
+    Oto = 2,
+    Ouo = 3,
 }
 
-impl From<&str> for AggregationSource {
-    fn from(s: &str) -> Self {
-        match s.to_uppercase().as_str() {
-            "EXTERNAL" => AggregationSource::EXTERNAL,
-            "INTERNAL" => AggregationSource::INTERNAL,
-            _ => panic!("Invalid `AggregationSource` value, was {s}"),
-        }
-    }
-}
-impl From<u8> for AggregationSource {
-    fn from(i: u8) -> Self {
-        match i {
-            1 => AggregationSource::EXTERNAL,
-            2 => AggregationSource::INTERNAL,
-            _ => panic!("Invalid `AggregationSource` value, was {i}"),
-        }
-    }
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum CurrencyType {
+    Crypto = 1,
+    Fiat = 2,
 }
 
-impl Display for AggregationSource {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(f, "{}", self.as_str())
-    }
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum DepthType {
+    Volume = 1,
+    Exposure = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum InstrumentCloseType {
+    EndOfSession = 1,
+    Expired = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum InstrumentStatus {
+    Closed = 1,
+    PreOpen = 2,
+    Open = 3,
+    Pause = 4,
+    PreClose = 5,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum LiquiditySide {
+    None = 0,
+    Maker = 1,
+    Taker = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum OmsType {
+    None = 0,
+    Netting = 1,
+    Hedging = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum OptionKind {
+    Call = 1,
+    Put = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum OrderSide {
+    None = 0,
+    Buy = 1,
+    Sell = 2,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum OrderStatus {
+    Initialized = 1,
+    Denied = 2,
+    Submitted = 3,
+    Accepted = 4,
+    Rejected = 5,
+    Canceled = 6,
+    Expired = 7,
+    Triggered = 8,
+    PendingUpdate = 9,
+    PendingCancel = 10,
+    PartiallyFilled = 11,
+    Filled = 12,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum OrderType {
+    Market = 1,
+    Limit = 2,
+    StopMarket = 3,
+    StopLimit = 4,
+    MarketToLimit = 5,
+    MarketIfTouched = 6,
+    LimitIfTouched = 7,
+    TrailingStopMarket = 8,
+    TrailingStopLimit = 9,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum PositionSide {
+    None = 0,
+    Flat = 1,
+    Long = 2,
+    Short = 3,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum PriceType {
+    Bid = 1,
+    Ask = 2,
+    Mid = 3,
+    Last = 4,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum TimeInForce {
+    Gtc = 1,
+    Ioc = 2,
+    Fok = 3,
+    Gtd = 4,
+    Day = 5,
+    AtTheOpen = 6,
+    AtTheClose = 7,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum TradingState {
+    Active = 1,
+    Halted = 2,
+    Reducing = 3,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum TrailingOffsetType {
+    None = 0,
+    Default = 1,
+    Price = 2,
+    BasisPoints = 3,
+    Ticks = 4,
+    PriceTier = 5,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum TriggerType {
+    None = 0,
+    Default = 1,
+    BidAsk = 2,
+    Last = 3,
+    DoubleLast = 4,
+    DoubleBidAsk = 5,
+    LastOrBidAsk = 6,
+    MidPoint = 7,
+    Mark = 8,
+    Index = 9,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, FromRepr, EnumString, Display)]
+#[strum(ascii_case_insensitive)]
+#[strum(serialize_all = "SCREAMING_SNAKE_CASE")]
+pub enum VenueStatus {
+    Closed = 1,
+    PreOpen = 2,
+    Open = 3,
+    Pause = 4,
+    PreClose = 5,
+}
+
+// TODO(cs): These should be macros
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes that since the data is originating from Rust, the GIL does not need
+/// to be acquired.
+/// - Assumes you are immediately returning this pointer to Python.
+#[no_mangle]
+pub unsafe extern "C" fn account_type_to_pystr(value: AccountType) -> *mut ffi::PyObject {
+    string_to_pystr(&value.to_string())
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
+#[no_mangle]
+pub unsafe extern "C" fn account_type_from_pystr(ptr: *mut ffi::PyObject) -> AccountType {
+    AccountType::from_str(&pystr_to_string(ptr)).unwrap()
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes that since the data is originating from Rust, the GIL does not need
+/// to be acquired.
+/// - Assumes you are immediately returning this pointer to Python.
+#[no_mangle]
+pub unsafe extern "C" fn aggregation_source_to_pystr(
+    value: AggregationSource,
+) -> *mut ffi::PyObject {
+    string_to_pystr(&value.to_string())
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
+#[no_mangle]
+pub unsafe extern "C" fn aggregation_source_from_pystr(
+    ptr: *mut ffi::PyObject,
+) -> AggregationSource {
+    AggregationSource::from_str(&pystr_to_string(ptr)).unwrap()
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes that since the data is originating from Rust, the GIL does not need
+/// to be acquired.
+/// - Assumes you are immediately returning this pointer to Python.
+#[no_mangle]
+pub unsafe extern "C" fn aggressor_side_to_pystr(value: AggressorSide) -> *mut ffi::PyObject {
+    string_to_pystr(&value.to_string())
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
+#[no_mangle]
+pub unsafe extern "C" fn aggressor_side_from_pystr(ptr: *mut ffi::PyObject) -> AggressorSide {
+    AggressorSide::from_str(&pystr_to_string(ptr)).unwrap()
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes that since the data is originating from Rust, the GIL does not need
+/// to be acquired.
+/// - Assumes you are immediately returning this pointer to Python.
+#[no_mangle]
+pub unsafe extern "C" fn asset_class_to_pystr(value: AssetClass) -> *mut ffi::PyObject {
+    string_to_pystr(&value.to_string())
+}
+
+/// Returns a pointer to a valid Python UTF-8 string.
+///
+/// # Safety
+/// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
+#[no_mangle]
+pub unsafe extern "C" fn asset_class_from_pystr(ptr: *mut ffi::PyObject) -> AssetClass {
+    AssetClass::from_str(&pystr_to_string(ptr)).unwrap()
 }