+++ nautilus_trader/execution/emulator.pyx
@@ -42,6 +42,7 @@ from nautilus_trader.execution.trailing cimport TrailingStopCalculator
 from nautilus_trader.model.c_enums.contingency_type cimport ContingencyType
 from nautilus_trader.model.c_enums.order_side cimport OrderSide
 from nautilus_trader.model.c_enums.order_type cimport OrderType
+from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
 from nautilus_trader.model.c_enums.trigger_type cimport TriggerType
 from nautilus_trader.model.c_enums.trigger_type cimport TriggerTypeParser
 from nautilus_trader.model.data.tick cimport QuoteTick
@@ -676,7 +677,9 @@ cdef class OrderEmulator(Actor):
                 ts_event=self._clock.timestamp_ns(),
                 ts_init=self._clock.timestamp_ns(),
             )
-            self._send_exec_event(event)
+            # TODO(cs): Determine a way of publishing event without applying
+            order.apply(event)
+            # self._send_exec_event(event)
             self._fill_limit_order(order)
         elif (
             order.order_type == OrderType.STOP_MARKET
@@ -761,7 +764,7 @@ cdef class OrderEmulator(Actor):
 
     cpdef void on_quote_tick(self, QuoteTick tick) except *:
         if not self._log.is_bypassed:
-            self._log.debug(f"Processing {repr(tick)}...")
+            self._log.debug(f"Processing {repr(tick)}...", LogColor.CYAN)
 
         cdef MatchingCore matching_core = self._matching_cores.get(tick.instrument_id)
         if matching_core is None:
@@ -775,7 +778,7 @@ cdef class OrderEmulator(Actor):
 
     cpdef void on_trade_tick(self, TradeTick tick) except *:
         if not self._log.is_bypassed:
-            self._log.debug(f"Processing {repr(tick)}...")
+            self._log.debug(f"Processing {repr(tick)}...", LogColor.CYAN)
 
         cdef MatchingCore matching_core = self._matching_cores.get(tick.instrument_id)
         if matching_core is None:
@@ -878,7 +881,7 @@ cdef class OrderEmulator(Actor):
             client_order_id=order.client_order_id,
             order_side=order.side,
             quantity=order.quantity,
-            time_in_force=order.time_in_force,
+            time_in_force=order.time_in_force if order.time_in_force != TimeInForce.GTD else TimeInForce.GTC,
             reduce_only=order.is_reduce_only,
             init_id=UUID4(),
             ts_init=self._clock.timestamp_ns(),