+++ nautilus_trader/model/enums.pyx
@@ -13,125 +13,11 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-"""
-Defines the enums of the trading domain model.
+"""Defines the enums of the trading domain model."""
 
-Account Type
-------------
-Represents a trading account type.
-
->>> from nautilus_trader.model.enums import AccountType
->>> AccountType.CASH
-<AccountType.CASH: 1>
->>> AccountType.MARGIN
-<AccountType.MARGIN: 2>
->>> AccountType.BETTING
-<AccountType.BETTING: 3>
-
-Aggregation Source
-------------------
-Represents where a bar was aggregated in relation to the platform.
-
->>> from nautilus_trader.model.enums import AggregationSource
->>> AggregationSource.EXTERNAL  # Bar was aggregated externally to the platform
-<AggregationSource.EXTERNAL: 1>
->>> AggregationSource.INTERNAL  # Bar was aggregated internally within the platform
-<AggregationSource.INTERNAL: 2>
-
-Aggregssor Side
----------------
-Represents the order side of the aggressor (liquidity taker) for a particular trade.
-
->>> from nautilus_trader.model.enums import AggressorSide
->>> AggressorSide.BUY
-<AggressorSide.BUY: 1>
->>> AggressorSide.SELL
-<AggressorSide.SELL: 2>
-
-Asset Class
------------
-Represents a group of investment vehicles with similar properties and risk profiles.
-
->>> from nautilus_trader.model.enums import AssetClass
->>> AssetClass.FX
-<AssetClass.FX: 1>
->>> AssetClass.EQUITY
-<AssetClass.EQUITY: 2>
->>> AssetClass.COMMODITY
-<AssetClass.COMMODITY: 3>
->>> AssetClass.METAL
-<AssetClass.METAL: 4>
->>> AssetClass.ENERGY
-<AssetClass.ENERGY: 5>
->>> AssetClass.BOND
-<AssetClass.BOND: 6>
->>> AssetClass.INDEX
-<AssetClass.INDEX: 7>
->>> AssetClass.CRYPTO
-<AssetClass.CRYPTO: 8>
->>> AssetClass.BETTING
-<AssetClass.BETTING: 9>
-
-Asset Type
-----------
-Represents a group of financial product types.
-
->>> from nautilus_trader.model.enums import AssetType
->>> AssetType.SPOT
-<AssetType.SPOT: 1>
->>> AssetType.SWAP
-<AssetType.SWAP: 2>
->>> AssetType.FUTURE
-<AssetType.FUTURE: 3>
->>> AssetType.FORWARD
-<AssetType.FORWARD: 4>
->>> AssetType.CFD
-<AssetType.CFD: 5>
->>> AssetType.OPTION
-<AssetType.OPTION: 6>
->>> AssetType.WARRANT
-<AssetType.WARRANT: 7>
-
-Bar Aggregation
----------------
-Represents a method of aggregating an OHLCV bar.
-
->>> from nautilus_trader.model.enums import BarAggregation
->>> BarAggregation.TICK
-<BarAggregation.TICK: 1>
->>> BarAggregation.TICK_IMBALANCE
-<BarAggregation.TICK_IMBALANCE: 2>
->>> BarAggregation.TICK_RUNS
-<BarAggregation.TICK_RUNS: 3>
->>> BarAggregation.VOLUME
-<BarAggregation.VOLUME: 4>
->>> BarAggregation.VOLUME_IMBALANCE
-<BarAggregation.VOLUME_IMBALANCE: 5>
->>> BarAggregation.VOLUME_RUNS
-<BarAggregation.VOLUME_RUNS: 6>
->>> BarAggregation.VALUE
-<BarAggregation.VALUE: 7>
->>> BarAggregation.VALUE_IMBALANCE
-<BarAggregation.VALUE_IMBALANCE: 8>
->>> BarAggregation.VALUE_RUNS
-<BarAggregation.VALUE_RUNS: 9>
->>> BarAggregation.SECOND
-<BarAggregation.SECOND: 10>
->>> BarAggregation.MINUTE
-<BarAggregation.MINUTE: 11>
->>> BarAggregation.HOUR
-<BarAggregation.HOUR: 12>
->>> BarAggregation.DAY
-<BarAggregation.DAY: 13>
-
-"""
-
-from nautilus_trader.model.c_enums.account_type import AccountType
-from nautilus_trader.model.c_enums.account_type import AccountTypeParser
-from nautilus_trader.model.c_enums.aggregation_source import AggregationSource
-from nautilus_trader.model.c_enums.aggregation_source import AggregationSourceParser
-from nautilus_trader.model.c_enums.aggressor_side import AggressorSide
-from nautilus_trader.model.c_enums.aggressor_side import AggressorSideParser
+from nautilus_trader.core.rust.enums import AggressorSide
+from nautilus_trader.core.rust.model import AccountType
+from nautilus_trader.core.rust.model import AggregationSource
 from nautilus_trader.model.c_enums.asset_class import AssetClass
 from nautilus_trader.model.c_enums.asset_class import AssetClassParser
 from nautilus_trader.model.c_enums.asset_type import AssetType
@@ -179,14 +65,18 @@ from nautilus_trader.model.c_enums.trigger_type import TriggerTypeParser
 from nautilus_trader.model.c_enums.venue_status import VenueStatus
 from nautilus_trader.model.c_enums.venue_status import VenueStatusParser
 
+from nautilus_trader.core.rust.enums cimport account_type_from_str
+from nautilus_trader.core.rust.enums cimport account_type_to_str
+from nautilus_trader.core.rust.enums cimport aggregation_source_from_str
+from nautilus_trader.core.rust.enums cimport aggregation_source_to_str
+from nautilus_trader.core.rust.enums cimport aggressor_side_from_str
+from nautilus_trader.core.rust.enums cimport aggressor_side_to_str
+
 
 __all__ = [
     "AccountType",
-    "AccountTypeParser",
     "AggregationSource",
-    "AggregationSourceParser",
     "AggressorSide",
-    "AggressorSideParser",
     "AssetClass",
     "AssetClassParser",
     "AssetType",
@@ -233,4 +123,10 @@ __all__ = [
     "TrailingOffsetTypeParser",
     "VenueStatus",
     "VenueStatusParser",
+    "account_type_to_str",
+    "account_type_from_str",
+    "aggregation_source_to_str",
+    "aggregation_source_from_str",
+    "aggressor_side_to_str",
+    "aggressor_side_from_str",
 ]