+++ RELEASES.md
@@ -1,3 +1,24 @@
+# NautilusTrader 1.164.0 Beta
+
+Released on 23rd December 2022 (UTC).
+
+### Breaking Changes
+None
+
+### Enhancements
+- Added managed GTD order expiry (experimental feature, config may change)
+- Added Rust `ParquetReader` and `ParquetWriter` (for `QuoteTick` and `TradeTick` only)
+
+### Fixes
+- Fixed `MARKET_IF_TOUCHED` orders for `OrderFactory.bracket(..)`
+- Fixed `OrderEmulator` trigger event handling for live trading
+- Fixed `OrderEmulator` transformation to market orders which had a GTD time in force
+- Fixed serialization of `OrderUpdated` events
+- Fixed typing and edge cases for new `msgspec`, thanks @limx0 
+- Fixed data wrangler processing with missing data, thanks @rsmb7z
+
+---
+
 # NautilusTrader 1.163.0 Beta
 
 Released on 17th December 2022 (UTC).